Migrate DeprecatedTooltip callsites to @ui/Tooltip and delete deprecated components
- Migrate all DeprecatedTooltip/Trigger/Content callsites to @ui/Tooltip
- Add className passthrough to TooltipTrigger for Layer__input-tooltip
- Move TooltipOptions type into useTooltip; delete deprecated Tooltip.tsx
- Delete styles/tooltip.scss and the Layer__tooltip selectors it backed
- Rename Layer__UI__tooltip* classes to PascalCase BEM (TooltipTrigger/TooltipContent/TooltipContent__Text)
- Drop external Layer__UI__TooltipTrigger style overrides
- Render the text span inside TooltipContent and consolidate tooltip.scss
