+++ src/components/DateSelection/globalDateRangeSelection.scss
@@ -11,8 +11,4 @@
       grid-column: 1 / -1;
     }
   }
-
-  .Layer__tooltip-trigger {
-    padding-top: 0;
-  }
 }
