+++ src/components/MatchForm/matchForm.scss
@@ -71,10 +71,6 @@
   overflow: hidden;
   cursor: pointer;
 
-  .Layer__tooltip-trigger {
-    padding-top: 0;
-  }
-
   & > div {
     display: flex;
     align-items: center;
@@ -207,10 +203,6 @@
   .Layer__MatchForm__Table__desc {
     gap: var(--spacing-sm);
     justify-content: space-between;
-
-    & > .Layer__UI__tooltip-trigger:first-child {
-      width: calc(100% - 220px);
-    }
 
     & > span:nth-child(2) {
       flex: 0;
