+++ src/styles/index.scss
@@ -17,7 +17,6 @@
 @forward './loader';
 @forward './profit_and_loss';
 @forward './profit_and_loss_detailed_charts';
-@forward './tooltip';
 @forward './accounting_overview';
 @forward './bookkeeping_overview';
 @forward './reports';
