--- src/styles/tooltip.scss
@@ -1,26 +1,0 @@
-.Layer__tooltip {
-  box-sizing: border-box;
-  z-index: 100;
-  width: max-content;
-  max-width: calc(100vw - 10px);
-
-  &[data-width='md'] {
-    max-width: 256px;
-  }
-}
-
-.Layer__tooltip .Layer__tooltip-content {
-  box-sizing: border-box;
-  padding: var(--spacing-2xs) var(--spacing-sm);
-  border-radius: var(--border-radius-2xs);
-  background: var(--color-base-1000);
-  font-size: var(--text-sm);
-  color: var(--color-base-300);
-}
-
-.Layer__tooltip-trigger {
-  display: flex;
-  flex: 1;
-  max-width: 100%;
-  padding-top: 2px;
-}
