+++ src/styles/variables.scss
@@ -2,8 +2,7 @@
 .Layer__Portal,
 .Layer__view,
 .Layer__select__menu-portal,
-.Layer__tooltip,
-.Layer__UI__tooltip,
+.Layer__UI__Tooltip,
 .Layer__toasts-container,
 .Layer__drawer,
 .Layer__variables,
@@ -209,8 +208,7 @@
   .Layer__Portal,
   .Layer__view,
   .Layer__select__menu-portal,
-  .Layer__tooltip,
-  .Layer__UI__tooltip,
+  .Layer__UI__Tooltip,
   .Layer__toasts-container,
   .Layer__drawer,
   .Layer__variables,
